Gestion de permissions, Les permissions du modèle Guide sont fonctionnelle

+++ app/Http/Controllers/Api/GuideController.php
@@ -5,18 +5,26 @@ namespace App\Http\Controllers\Api;
 use App\Models\Guide;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Traits\ReturnJsonResponseTrait;
+use Exception;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class GuideController extends Controller
 {
+    use ReturnJsonResponseTrait;
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Guide $guide)
     {
-        $guides = Guide::all();
-        return response()->json(compact('guides'),200);
+        $this->authorize('view', $guide);
+        try {
+            $guides = Guide::all();
+            return response()->json(compact('guide'), 200);
+        } catch (Exception $error) {
+            return response()->json(['message' => $error->getMessage()]);
+        }
     }
 
     /**
@@ -45,31 +53,49 @@ class GuideController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request, Guide $guide)
     {
-       $user = Auth::user();
-       $validator = Validator::make($request->all(), $this->rules());
 
-        if ($validator->fails()) {
-            return response()->json(['errors' => $validator->errors()], 422);
+        //$this->authorize('store', $guide);
+        try {
+
+            $user = Auth::user();
+            $validator = Validator::make($request->all(), $this->rules());
+
+            if ($validator->fails()) {
+                return response()->json(['errors' => $validator->errors()], 422);
+            }
+
+            $existeAutreGuide = Guide::all()->count();
+            if ($existeAutreGuide) {
+                return response()->json(['errors' => 'Vous ne pouvez ajouter qu\'un seul guide pour l\'instant '], 422);
+            }
+
+            $guide = new Guide();
+            $guide->titre = $request->titre;
+            $guide->description = $request->description;
+            $guide->user_id = $user->id;
+            $guide->save();
+
+            return response()->json(['message' => 'Guide créé avec succès'], 201);
+        } catch (Exception $error) {
+            return response()->json(['message' => $error->getMessage()]);
         }
-
-        $guide = new Guide();
-        $guide->titre = $request->titre;
-        $guide->description = $request->description;
-        $guide->user_id = $user->id;
-        $guide->save();
-
-        return response()->json(['message' => 'Guide créé avec succès'], 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show($id)
+    public function show($id, Guide $guide)
     {
-        $guide = Guide::find($id);
-        return response()->json(compact('guide'),200);
+        $this->authorize('view', $guide);
+
+        try {
+            $guide = Guide::find($id);
+            return response()->json(compact('guide'), 200);
+        } catch (Exception $error) {
+            return response()->json(['message' => $error->getMessage()]);
+        }
     }
 
     /**
@@ -83,30 +109,48 @@ class GuideController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, $id, Guide $guide)
     {
-       $user = Auth::user();
-       $validator = Validator::make($request->all(), $this->rules(), $this->messages());
 
-        if ($validator->fails()) {
-            return response()->json(['errors' => $validator->errors()], 422);
+        $this->authorize('update', $guide);
+        try {
+            $guideUpdated = Guide::findOrFail($request->id);
+           
+        } catch (Exception $error) {
+            return response()->json(['message' => 'Enregistrement innexistant']);
         }
 
-        $guide = Guide::find($id);
-        $guide->titre = $request->titre;
-        $guide->description = $request->description;
-        $guide->user_id = $user->id;
-        $guide->save();
+        try {
+           
+            $user = Auth::user();
+            $validator = Validator::make($request->all(), $this->rules(), $this->messages());
 
-        return response()->json(['message' => 'Guide modifié avec succès'], 201);
+            if ($validator->fails()) {
+                return response()->json(['errors' => $validator->errors()], 422);
+            }
+
+            $guideUpdated->titre = $request->titre;
+            $guideUpdated->description = $request->description;
+            $guideUpdated->user_id = $user->id;
+            $guideUpdated->save();
+            return response()->json(['message' => 'Guide modifié avec succès'], 201);
+        } catch (Exception $error) {
+            return response()->json(['message' => $error->getMessage()]);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy($id, Guide $guide)
     {
-        Guide::find($id)->delete();
-        return response()->json(['message' => 'Guide supprimé avec succès'], 200);
+        $this->authorize('delete', $guide);
+
+        try {
+            Guide::find($id)->delete();
+            return response()->json(['message' => 'Guide supprimé avec succès'], 200);
+        } catch (Exception $error) {
+            return response()->json(['message' => $error->getMessage()]);
+        }
     }
 }
